refactor: collapse the connection types into a single Xtdb.Connection
XTDB had two near-identical connection types, one nested inside the other: `xtdb.NodeConnection`, the connection-scoped handle that threads read-your-writes (await tokens), and `xtdb.adbc.XtdbConnection`, an `AdbcConnection` that *wrapped* that handle and was mostly pure delegation — `prepareSql`/`openSqlQuery`/`openSnapshot`/`executeTx` forwarded straight through, `setCurrentCatalog` mutated the handle's `dbName`, and a couple of helpers re-exposed what the handle already offered.

Collapse them into one nested `Xtdb.Connection` that *is* an `AdbcConnection` — the handle and the adapter become the same object. pgwire uses its XT-native surface (await-token-threaded `submitTx`/`openSqlQuery`/…); ADBC and Flight SQL use the `AdbcConnection` surface on that same connection. The ADBC statement nests alongside as `Xtdb.Statement`, and the two per-frontend `Node` SPIs merge into one, so the node impl exposes a single connection factory feeding both `connect` (ADBC/Flight) and `open-connection` (pgwire).

Narrowing `connect()` to return `Xtdb.Connection` lets the `as XtdbConnection` casts in the Flight SQL producer, the TPC-H loader and the ADBC test go away.

No runtime behaviour changes — it's a structural collapse. One thing is deliberately left: a pgwire-driven connection still carries the ADBC transaction buffer (`autoCommit`/`pendingOps`) it never touches, since pgwire runs its own transaction state. Unifying those is the open question behind a later pgwire migration, out of scope here.
